Handle forked children in tests
The trick is to not trigger the TAP summary/exit behavior when the child
process (who inherits the atexit() handlers) fails to exec and exits
abnormally.  This is done by tracking the PID who initiated the test
plan and skipping the eval_test_run() call in atexit() unless the
exiting process' PID matches that.